[fem] Prepare to move utility methods out of dev
* Move implementation to .cc files.
* Move PermuteBlockSparseMatrix() to matrix_utilities.h
* Rename constitutive_model_utilities.h to calc_lame_parameters.h since
  it only has a single method.
